CORE-1181 Migrate tokens SDK project to build using the corda 5 API  (#1)
Updated project to build using latest 5.0-DevPreview1 branch of corda5. Latest DevPreview of corda5 will be release as RC03. This included upgrading to Java 11, Kotlin 1.4, Gradle 6.8.3, and the latest corda gradle plugin to allow cordapps to build as CPKs. The target release number has been incremented from 1.0 to 2.0 since it is building against a major number increment of corda.

Remaining work will be to:
* Set the target platform version for cordapp configurations (https://r3-cev.atlassian.net/browse/CORE-1374)
* Migrate the old DriverDSL/MockNetwork tests: https://r3-cev.atlassian.net/browse/CORE-1373

The existing tests have been removed in this PR to be replaced later. Proper unit tests will be added in a follow on PR.