CORE-2298 Reintroduce infix notation for building tokens. (#5)
The infix functions reported as missing in the bug were lost because these infix functions initialised the token objects, and when updating to the latest C5 API a HashingService became part of the constructor for these token objects. The infix functions couldn't allow for multiple parameters to support this so these functions lost infix when updated to take multiple params.

These PR reintroduces the removed infix functions and introduces new infix functions withHashingService. To allow the tokens to be built across multiple infix functions, they were updated to be backed by a builder class. The builder classes already existed but used the infix functions. This has been flipped so that the infix functions use the builder classes instead.

The location of utility classes for build fungible and non fungible tokens was inconsistent so all have been moved together in a new tokens-builder module. Previously the fungible token builder utils were in the contracts module just so it could be used by both selection and workflows, and non-fungible token builder utils were in the workflows module. The contracts module wasn't a good location for these common classes so that's why the new module was introduced. Builder utils from workflows were moved for the sake of consistency.

Introduction of this new module does require it to be built as a cordapp CPK because it used corda jars. If it wasn't a CPK then when it's included as a dependency of other CPKs there is an error when starting the node saying the corda jars cannot be included in the CPKs. This means tokens SDK produces 4 CPKs now; contracts, workflows, selection and tokens-builder.

The builder classes previously built Amount<IssuedTokenType> or IssuedTokenType, but to allow the infix utility functions to work the builders have been updated to accept these objects pre-built.

Corda 4:
val fungibleToken: FungibleToken = amount of ccy issuedBy ourIdentity heldBy issueTo

Corda 5:
val fungibleToken: FungibleToken = amount of ccy issuedBy ourIdentity heldBy issueTo withHashingService hashingService